Remove unused testconstants imports and fix type mismatch
  - Remove testconstants import from test files where it's no longer used
  - Fix TestRequireSameWEVMOSMainnet by casting config.ShardeumChainID to uint64

+++ testutil/constants/constants_test.go
@@ -28,7 +28,7 @@ func TestRequireSameTestBech32Prefix(t *testing.T) {
 func TestRequireSameWEVMOSMainnet(t *testing.T) {
 	require.Equal(t,
 		constants.ShardeumChainID.EVMChainID,
-		config.ShardeumChainID,
+		uint64(config.ShardeumChainID),
 		"EVM chain IDs should be the same across the repo",
 	)
 }
